backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.8.0
Flask-Cors==4.0.1
gunicorn==22.0.0 